New concurrent design using sync blocks
MVar created problems with for each loops, therefore it had to be
removed as it would stop the loop from functioning as it was blocked out
by threads.

The new design uses sync blocks that use the trick as an object. Threads
are running correctly however better communication is needed for taking
results from trick.