Debugged the sync block design
You are able to run a game with 4AI players now, however errors occur
when the players are on their final cards. The score also needs
debugging as it can go up to 9038547... Other than that the design has
been sucessful, the table is given less tasks to do and admin work is
delagated onto the players.